backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.122.0",
    "motor>=3.7.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
